renaming validation file for customEvent
